requirements.txt
requests
pandas
jupyter
notebook
ipykernel
elasticsearch==7.17
pymongo
python-dotenv
bcrypt
flask